LPBUG#882689 - crash during startup on XP.
The reason for the crash is Innodb assertion after trying to load condition variables function
dynamically and not finding them

The fix is to skip dynamic loading if srv_use_native_conditions is FALSE. srv_use_native_conditions 
is derived from Windows version and would be FALSE on XP and TRUE on later Windows.

This is the same handling as in MySQL 5.. In Maria 5.3 srv_use_native_conditions check was 
presumably  lost in the downporting.